+++ storch/utils/layer_helpers.py
@@ -2,6 +2,13 @@
 from collections.abc import Iterable
 from itertools import repeat
 
+__all__=[
+    'to_1tuple',
+    'to_2tuple',
+    'to_3tuple',
+    'to_4tuple',
+]
+
 
 def _ntuple(n):
